src/pyproject.toml
[project]
name = "django-docker-template"
version = "2.0.0"
description = "Production-ready Django Docker boilerplate"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2,<6.1",
    "djangorestframework>=3.16,<4",
    "djangorestframework-simplejwt>=5.5,<6",
    "django-prometheus>=2.3,<3",
    "environs[django]>=14.1,<15",
    "psycopg[binary]>=3.2,<4",
    "gunicorn>=23,<24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-django>=4.9,<5",
    "factory-boy>=3.3,<4",
    "ruff>=0.9,<1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "main.settings"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py314"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["main", "accounts", "health", "logging_utils"]
